Make `dandi validate` support Zarrs
